requirements.txt
pytest==6.2.5
selenium==4.1.0
webdriver-manager==3.5.2
PyHamcrest